Update setVersions.sh with updates to fix the CI/CD errors
Update setVersions.sh with updates to cut the Update setVersions.sh with updates to cut "$VERSION_FILE with 8 characters

+++ deploy/setVersions.sh
@@ -1,23 +1,59 @@
 #!/bin/bash
+set -ex
 
 VERSION_FILE="$HOME/deploy/versions.json"
 ENV_SAMPLE_FILE="$HOME/deploy/.env.sample"
 ENV_FILE="$HOME/app/.env"
 
-# Extract versions from json
-FE_HASH=$(jq -r '."ERAS-FE"' "$VERSION_FILE" | cut -c1-7)
+# Validar existencia de archivos
+test -f "$VERSION_FILE" || (echo "Error: $VERSION_FILE not found" && exit 1)
+test -f "$ENV_SAMPLE_FILE" || (echo "Error: $ENV_SAMPLE_FILE not found" && exit 1)
+
+# Mostrar contenido del JSON
+echo "Contenido de $VERSION_FILE:"
+cat "$VERSION_FILE"
+
+# Extraer versiones
+FE_HASH=$(jq -r '."ERAS-FE"' "$VERSION_FILE" | cut -c1-8)
 BE_HASH=$(jq -r '."ERAS-BE"' "$VERSION_FILE" | cut -c1-7)
 
-# Verify if versions were succesfully extracted
+# Extraer environment
+FE_ENV=$(jq -r '."ERAS-FE"' "$VERSION_FILE" | cut -c10-)
+BE_ENV=$(jq -r '."ERAS-BE"' "$VERSION_FILE" | cut -c9-)
+
+
+echo "FE_HASH: $FE_HASH"
+echo "BE_HASH: $BE_HASH"
+
+# Verificar extracción
 if [[ -z "$FE_HASH" || -z "$BE_HASH" ]]; then
-    echo "Error: Couldn't extract versions from  $VERSION_FILE"
+    echo "Error: Couldn't extract versions from $VERSION_FILE"
     exit 1
 fi
 
+# Copiar archivo base
 cp "$ENV_SAMPLE_FILE" "$ENV_FILE"
 
-# Replace versions values at new .env file
-sed -i "s/_ERAS-FE_/$FE_HASH/" "$ENV_FILE"
-sed -i "s/_ERAS-BE_/$BE_HASH/" "$ENV_FILE"
 
-echo "File .env generated - FE_VERSION=0.1.0-$FE_HASH-dev & BE_VERSION=0.1.0-$BE_HASH-dev"
+# Reemplazar versiones FE
+if [ "$FE_ENV" != "dev" ]; then
+  # Si el entorno NO es dev → reemplaza con hash + entorno y elimina el -dev final
+  sed -i "s/_ERAS-FE_/${FE_HASH}-${FE_ENV}/" "$ENV_FILE"
+  sed -i '/FE_VERSION/s/-dev$//' "$ENV_FILE"
+else
+  # Si el entorno ES dev → solo reemplaza el hash
+  sed -i "s/_ERAS-FE_/${FE_HASH}/" "$ENV_FILE"
+fi
+
+# Reemplazar versiones BE
+if [ "$BE_ENV" != "dev" ]; then
+  sed -i "s/_ERAS-BE_/${BE_HASH}-${BE_ENV}/" "$ENV_FILE"
+  sed -i '/BE_VERSION/s/-dev$//' "$ENV_FILE"
+else
+  sed -i "s/_ERAS-BE_/${BE_HASH}/" "$ENV_FILE"
+fi
+
+
+# Mostrar resultado final
+echo "Archivo .env generado:"
+cat "$ENV_FILE"
